- Fix unittest - Update readme - Update TODO - Pin Pillow to lower version 9.1.0 - Update requires to major than version instead of static version
