+++ CHANGELOG.txt
@@ -1,3 +1,10 @@
+3.0.7
+-----
+- Fix unittest
+- Start dropping Python 2 support
+- Pin dependency Pillow to a lower version 9.1.0
+- Fix demo
+
 3.0.6
 -----
 - Fix compatibility with Pillow 10.0.0 (ANTIALIAS is deprecated, using LANCZOS instead)
