+++ setup.py
@@ -15,10 +15,10 @@ setup(
 
     install_requires=[
         'Flask>=0.9',
-        'itsdangerous==2.1.2',  # For Flask v0.9
-        'Pillow==10.2.0',
-        'six==1.16.0',
-        'scour==0.38.2',
+        'itsdangerous>=2.1.2',  # For Flask v0.9
+        'Pillow>=9.1.0',
+        'six>=1.16.0',
+        'scour>=0.38.2',
     ],
 
     classifiers=[
